refactor(statements): delete unused v1 build code (#580)
Now that statements v2 — including the Polygon export/upload path (#568) —
owns every live build, the dangling v1 machinery is dead weight. Remove it:

- builders.py: the StatementBuilder classes + builder-chain. The 5 still-live
  low-level helpers (StatementBlocks, render_jinja, render_jinja_blocks,
  externalize_blocks, substitute_externalized_blocks) move into render.py.
- joiners.py: v1 contest-join builders (v2 joins via \subimport).
- statement_utils.py (+ test): get_relative_assets, unused since the overlay
  mirrors whole asset dirs.
- build_statements.py: the dead v1 builder-chain resolver (get_builders et al.).
- schema.py: dead conversion/joiner models (JinjaTeX, rbxMarkdownToTeX,
  JoinerType, JoinTexToPDF, Joiner). The export-time externalize/demacro
  vocabulary (ConversionType/ConversionStep/rbxToTeX/TexToPDF) stays.

Net -1161 lines, no behavior change.
